Hide And Sea Passive Event

+++ Assets/Scripts/Game/Events/EventTiming.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public struct EventTiming
 {
 	[Tooltip("How often does the event repeat")]
-    public float Frequency;
+    public uint Frequency;
 	[Tooltip("Time before event can trigger again")]
 	public float Cooldown;
 	[Tooltip("How long does the event last")]
